+++ Sources/Conduit/Networking/Serialization/FormEncodedRequestSerializer.swift
@@ -14,7 +14,7 @@ public final class FormEncodedRequestSerializer: HTTPRequestSerializer {
     /// Defines how parameters should be encoded within the HTTP body.
     public var formattingOptions = QueryStringFormattingOptions()
 
-    override open func serialize(request: URLRequest, bodyParameters: Any? = nil) throws -> URLRequest {
+    override public func serialize(request: URLRequest, bodyParameters: Any? = nil) throws -> URLRequest {
 
         var mutableRequest = try super.serialize(request: request, bodyParameters: bodyParameters)