+++ Sources/Conduit/Networking/Serialization/XML/XML.swift
@@ -14,7 +14,7 @@ public enum XMLError: Error {
 }
 
 /// Represents an XML document
-public struct XML {
+public final class XML {
 
     /// The root of the document
     public var root: XMLNode?
@@ -40,12 +40,7 @@ extension XML: CustomStringConvertible {
 
     /// Serialized XML string output
     public var description: String {
-        var nodes = [XMLNode.versionInstruction]
-        nodes.append(contentsOf: processingInstructions)
-        if let root = root {
-            nodes.append(root)
-        }
-        return nodes.map { $0.description }.joined()
+        return xmlString(format: .condensed)
     }
 
 }
@@ -53,18 +48,16 @@ extension XML: CustomStringConvertible {
 // MARK: - LosslessStringConvertible
 
 extension XML: LosslessStringConvertible {
-
     /// Attempts to produce an XML document with the provided XML string
     ///
     /// - Parameter description: The XML string to deserialize
-    public init?(_ description: String) {
+    public convenience init?(_ description: String) {
         let parser = XML.Parser(xmlString: description)
-        guard let xml = parser?.parse() else {
+        guard let xml = parser?.parse(), let root = xml.root else {
             return nil
         }
-        self = xml
+        self.init(root: root)
     }
-
 }
 
 // MARK: XML string parser
@@ -87,8 +80,7 @@ extension XML {
         }
 
         func parse() -> XML? {
-            if xmlParser.parse(),
-                let root = root {
+            if xmlParser.parse(), let root = root {
                 return XML(root: root)
             }
             return nil
@@ -96,8 +88,8 @@ extension XML {
 
         func parser(_ parser: XMLParser, didStartElement elementName: String, namespaceURI: String?, qualifiedName qName: String?,
                     attributes attributeDict: [String: String] = [:]) {
-            var node = XMLNode(name: elementName)
-            node.attributes = attributeDict
+            let attributes = XMLNodeAttributes(attributes: attributeDict.map { (attribute: $0.key, value: $0.value) })
+            let node = XMLNode(name: elementName, attributes: attributes)
             if let parentNode = workingTree.popLast() {
                 workingTree.append(parentNode)
             }
@@ -105,7 +97,7 @@ extension XML {
         }
 
         func parser(_ parser: XMLParser, foundCharacters string: String) {
-            if var activeNode = workingTree.popLast() {
+            if let activeNode = workingTree.popLast() {
                 let text = activeNode.text ?? ""
                 activeNode.text = text + string
                 workingTree.append(activeNode)
@@ -115,7 +107,7 @@ extension XML {
         func parser(_ parser: XMLParser, didEndElement elementName: String, namespaceURI: String?, qualifiedName qName: String?) {
             let finishedNode = workingTree.popLast()
             let parentNode = workingTree.popLast()
-            if var parentNode = parentNode {
+            if let parentNode = parentNode {
                 if let finishedNode = finishedNode {
                     var children = parentNode.children
                     children.append(finishedNode)