+++ Sources/Conduit/Networking/Serialization/XML/XMLNode.swift
@@ -12,36 +12,50 @@ import Foundation
 public typealias XMLDictionary = [String: CustomStringConvertible]
 
 /// Represents a single node in an XML document
-public struct XMLNode {
+public final class XMLNode {
 
     /// Not technically a PI, but it follows the same formatting rules
     static var versionInstruction: XMLNode = {
         var node = XMLNode(name: "xml")
-        node.attributes = [
-            "version": "1.0",
-            "encoding": "utf-8"
-        ]
+        node.attributes["version"] = "1.0"
+        node.attributes["encoding"] = "utf-8"
         node.isProcessingInstruction = true
         return node
     }()
 
     /// The name of the element
     public var name: String
+
     /// The child nodes
-    public var children: [XMLNode]
-    /// The element attributes
-    public var attributes = [String: String]()
+    public var children: [XMLNode] {
+        didSet {
+            resetChildrenParent()
+        }
+    }
+
+    /// Parent node. Set when node is added as children to another node.
+    public weak var parent: XMLNode?
+
+    /// Parent nodes all the way to the root node in the XML tree.
+    public var parents: [XMLNode] {
+        guard let parent = parent else {
+            return []
+        }
+        return [parent] + parent.parents
+    }
+
+    /// Collection of element attributes (preserves order)
+    public var attributes = XMLNodeAttributes()
+
     /// The contained text node (value) of the element
     public var text: String?
+
     /// Determines whether or not the element is a processing instruction
     public var isProcessingInstruction = false
+
     /// Determines whether or not the element is a leaf (no children)
     public var isLeaf: Bool {
         return children.isEmpty
-    }
-
-    private var isEmpty: Bool {
-        return isLeaf && text == nil
     }
 
     /// Construct XMLNode with optional value, attributes and children
@@ -51,11 +65,12 @@ public struct XMLNode {
     ///   - value: String value (text node)
     ///   - attributes: Node attributes dictionary
     ///   - children: Array of child nodes
-    public init(name: String, value: CustomStringConvertible? = nil, attributes: [String: String] = [:], children: [XMLNode] = []) {
+    public init(name: String, value: CustomStringConvertible? = nil, attributes: XMLNodeAttributes = XMLNodeAttributes(), children: [XMLNode] = []) {
         self.name = name
         self.text = value?.description
         self.attributes = attributes
         self.children = children
+        resetChildrenParent()
     }
 
     /// Construct XML node from a dictionary of [String: CustomStringConvertible]
@@ -83,6 +98,32 @@ public struct XMLNode {
             }
             return XMLNode(name: key, value: String(describing: value))
         }
+        resetChildrenParent()
+    }
+
+    private func resetChildrenParent() {
+        children.forEach { $0.parent = self }
+    }
+
+    /// Retrieve a list of all descendant nodes matching the given function
+    ///
+    /// - Parameter matching: Matching function to apply to each node
+    /// - Parameter traversal: Node Traversal technique.
+    /// - Returns: Array of descendant nodes
+    public func nodes(matching function: (XMLNode) -> Bool, traversal: XMLNodeTraversal) -> [XMLNode] {
+        if isLeaf {
+            return []
+        }
+
+        let matches = children.filter { function($0) }
+        if traversal == .firstLevel {
+            return matches
+        }
+
+        let breadth = traversal == .breadthFirst ? matches : []
+        let descendants = children.flatMap { $0.nodes(matching: function, traversal: traversal) }
+        let depth = traversal == .depthFirst ? matches : []
+        return breadth + descendants + depth
     }
 
     /// Retrieve a list of all descendant nodes with the given name
@@ -91,19 +132,7 @@ public struct XMLNode {
     /// - Parameter traversal: Node Traversal technique.
     /// - Returns: Array of descendant nodes
     public func nodes(named name: String, traversal: XMLNodeTraversal) -> [XMLNode] {
-        if isLeaf {
-            return []
-        }
-
-        let matches = children.filter { $0.name == name }
-        if traversal == .firstLevel {
-            return matches
-        }
-
-        let breadth = traversal == .breadthFirst ? matches : []
-        let descendants = children.flatMap { $0.nodes(named: name, traversal: traversal) }
-        let depth = traversal == .depthFirst ? matches : []
-        return breadth + descendants + depth
+        return nodes(matching: { $0.name == name }, traversal: traversal)
     }
 
     /// Retrieve the first descendant node with the given name
@@ -140,36 +169,7 @@ extension XMLNode: CustomStringConvertible {
 
     /// Serialized XML string output
     public var description: String {
-        let leftDelimiter = isProcessingInstruction ? "<?" : "<"
-        let rightDelimiter = isProcessingInstruction ? "?>" : (isEmpty ? "/>" : ">")
-
-        let hasAttributes = attributes.isEmpty == false
-        let startTag: String
-
-        if hasAttributes {
-            let describedAttributes = attributes.map { "\($0.key)=\"\($0.value)\"" }.joined(separator: " ")
-            startTag = "\(leftDelimiter)\(name) \(describedAttributes)\(rightDelimiter)"
-        }
-        else {
-            startTag = "\(leftDelimiter)\(name)\(rightDelimiter)"
-        }
-
-        let endTag = "\(leftDelimiter)/\(name)\(rightDelimiter)"
-
-        if isProcessingInstruction || isEmpty {
-            return startTag
-        }
-
-        if children.isEmpty == false {
-            let body = children.map { $0.description }.joined()
-            return "\(startTag)\(body)\(endTag)"
-        }
-
-        if let value = text {
-            return "\(startTag)\(value)\(endTag)"
-        }
-
-        return startTag
+        return xmlString(format: .condensed)
     }
 
 }
@@ -181,11 +181,11 @@ extension XMLNode: LosslessStringConvertible {
     /// Attempts to produce an XMLNode with the provided XML string
     ///
     /// - Parameter description: The XML string to deserialize
-    public init?(_ description: String) {
-        guard let node = XML(description)?.root else {
+    public convenience init?(_ description: String) {
+        guard let root = XML(description)?.root else {
             return nil
         }
-        self = node
+        self.init(name: root.name, value: root.getValue() as String?, attributes: root.attributes, children: root.children)
     }
 
 }